Try connecting to all possible URL paths
With multipage apps, we may have pages deployed to URLs like

* localhost:8501/my/base/path/streamlit_app.py
* localhost:8501/my/base/path/nested_dir/another_streamlit_app.py

which means that we can't just assume that all of our server URLs should
be relative to window.location. For example, the healthz endpoint for
this streamlit server will be localhost:8501/my/base/path/healthz

To handle this, we simply have the client try connecting to all possible
paths until it finds the base path. Thankfully, lots of the infrastructure
for this has already been built, which made writing this a whole lot less
work than I initially expected it to be.

We may want to redesign this to be more clever and less brute-force-y in
the hardened version of this feature, but for now just trying every
possible path works well enough.